docs/archive/oracle-pipeline/requirements.txt
anthropic>=0.40.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
praw>=7.8.0          # Reddit API
python-dotenv>=1.0.0
click>=8.1.0         # CLI
rich>=13.0.0         # Pretty terminal output
tenacity>=9.0.0      # Retry logic
chromadb>=0.5.0      # Vector store for knowledge artifact content
pyyaml>=6.0.0        # Vertical config files
